// src/app/shared/models/data-extraction.model.ts
export class DataExtraction {
  dataExtractionId: number;
  partInfoId: number;
  dataJson: string;
  cutoutPerimeterTotal: string;
  partInfo: any;
  bendLength: string;
  materialInfoJson?: string;
  processInfoJson?: string;
  cotsInfoJson?: string;
  coreInfoJson?: string;
  dfmIssues?: string;
  datumCentroid?: string;
}